requirements.txt
playwright
python-dotenv
pandas
rich
openpyxl
anthropic
tiktoken
